requirements.txt
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
requests>=2.31.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
lxml>=4.9.0
